fix: read customClaudeCodePath from ai-settings store in path loader
The customClaudeCodePathLoader in index.ts was calling store.get() which
reads from app-settings.json, but the value is saved by AIService to
ai-settings.json. This caused the loader to always return empty string,
falling back to the bundled SDK binary instead of the configured custom
executable.

Moves the loader setup into AIService constructor where it has direct
access to the ai-settings store via getSettingsStore().
